redis 설정 추가

+++ miniBacktesting_be/src/main/java/com/project/minibacktesting_be/dto/community/CommunityPortDto.java
@@ -15,7 +15,7 @@ public class CommunityPortDto {
     List<String> stockList;
     List<Integer> ratioList;
     Double finalYield;
-    Integer finalYieldMoney;
+    Double finalYieldMoney;
     Long seedMoney;
